Create SmartCrack module
